Add Home Page
I added a home page with a "Buy Now" button that floats in. I also
updated the CSS to work better with the background color. Finally, I
put message on the shop page that it is under progress, and made a dev
page that we can use for Stripe test mode.